requirements.txt
apikey==0.2.4
boto==2.49.0
boto3==1.18.26
botocore==1.21.26
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
jmespath==0.10.0
newsapi-python==0.2.6
oauthlib==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
schedule==1.1.0
six==1.16.0
tweepy==3.10.0
urllib3==1.26.6
